agent: Move `Tracker` to own file
